Add files via upload
As this repository is public the Supabase URL and Key in this file are secret. So not working as is. Replace those with your own URL and Key to get this working.

A function get_random_question on Supabase is called. This gets a random question from a table 'quiz_questions'.
I have to make this more generioc so you can send a parameter defining what function to call and what table will be targeted.